type conversion

+++ variables.go
@@ -81,10 +81,10 @@ var doDirectChipKick bool = false
 var doDirectKick bool = false
 
 type ImageData struct {
-	Is_ball_exit bool   `json:"isball"`
-	Image_x      uint32 `json:"image_x"`
-	Image_y      uint32 `json:"image_y"`
-	Frame        string `json:"frame"`
+	Is_ball_exit bool    `json:"isball"`
+	Image_x      float32 `json:"image_x"`
+	Image_y      float32 `json:"image_y"`
+	Frame        string  `json:"frame"`
 }
 
 var imageData ImageData
